- add the ability to send set requests to vinculum
